+++ scripts/supplier_parts.py
@@ -627,16 +627,21 @@ def supplier_parts(master_files, path):
 
     def supplier_parts_discontinued(cell_row, cell_col):
         part_no_supplier_code = PRIMARY_KEY_1 + PRIMARY_KEY_2
+        discontinue_list = []
 
         for row in range(9, selected['backup_0'].sheet_by_index(0).nrows):
             if part_no_supplier_code == str(selected['backup_0'].sheet_by_index(0).cell_value(row, 3)) + selected['backup_0'].sheet_by_index(0).cell_value(row, 11):
-                if selected['backup_0'].sheet_by_index(0).cell_value(row, 8) == 'N':
-                    print ('Discontinued check --- Pass')
-                    update_df('MOD', columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'PASS', part_no_supplier_code, selected['backup_0'].sheet_by_index(0).cell_value(row, 8), 'Part has not been discontinued')
-                else:
-                    print ('Discontinued check --- Fail')
-                    update_df('MOD', columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', part_no_supplier_code, selected['backup_0'].sheet_by_index(0).cell_value(row, 8), 'Part has already been discontinued')
+                discontinue_list.append((
+                    selected['backup_0'].sheet_by_index(0).cell_value(row, 5),
+                    selected['backup_0'].sheet_by_index(0).cell_value(row, 8)
+                ))
 
+        if 'N' in [tuple[1] for tuple in discontinue_list]:
+            print ('Discontinued check --- Pass')
+            update_df('MOD', columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'PASS', part_no_supplier_code, discontinue_list, 'Part has contract that has not been discontinued')
+        else:
+            print ('Discontinued check --- Fail')
+            update_df('MOD', columns[cell_col], cell_row, PRIMARY_KEY_1, PRIMARY_KEY_2, 'FAIL', part_no_supplier_code, discontinue_list, 'Part does not have contract that has not been discontinued')
 
     # Print required masters for checking